requirements.txt
python-pptx
requests
python-dotenv
rich
jsonschema
